requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
mediapipe>=0.10.0
huggingface_hub>=0.20.0
httpx>=0.24.0
